findOrCreateTestResult Service
